Curvature drive done

+++ car/include/motors.h
@@ -13,7 +13,7 @@ volatile uint16_t left_motor_pwm;
 volatile uint16_t right_motor_pwm;
 
 void motor_set(MotorSide side, uint32_t pct, MotorDir dir);
-void motor_set_bidir(MotorSide side, uint32_t pct_bi);
+void motor_set_bidir(MotorSide side, float pct_bi);
 void motors_set(uint32_t pct, MotorDir dir);
 void motor_forward(MotorSide side, uint32_t pct);
 void motors_forward(uint32_t pct);
